requirements.txt
openai-whisper
ffmpeg-python
streamlit
streamlit-audiorec
setuptools-rust